LayoutTests/media/media-source/media-source-webm-init-inside-segment.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-webm</title>
    <script src="media-source-loader.js"></script>
    <script src="../video-test.js"></script>
    <script>
    var loader;
    var source;
    var sourceBuffer;

    function loaderPromise(loader) {
        return new Promise((resolve, reject) => {
            loader.onload = resolve;
            loader.onerror = reject;
        });
    }

    window.addEventListener('load', async event => {
        try {
            findMediaElement();
            loader = new MediaSourceLoader('content/test-vp9-long-manifest.json');
            await loaderPromise(loader);

            source = new MediaSource();
            run('video.src = URL.createObjectURL(source)');
            await waitFor(source, 'sourceopen');
            waitFor(source, 'error').then(endTest);
            waitForEventAndFail('error');

            run('source.duration = loader.duration()');
            run('sourceBuffer = source.addSourceBuffer(loader.type())');
            run('sourceBuffer.appendBuffer(loader.initSegment())');

            await waitFor(sourceBuffer, 'update');

            consoleWrite('Append a media segment.')
            run('sourceBuffer.appendBuffer(loader.mediaSegment(0))');

            await waitFor(sourceBuffer, 'update');

            consoleWrite('Append a new Initialization Segment.')
            run('sourceBuffer.appendBuffer(loader.initSegment())');

            await waitFor(sourceBuffer, 'update');

            consoleWrite('Append a media segment.')
            run('sourceBuffer.appendBuffer(loader.mediaSegment(1))');

            await waitFor(sourceBuffer, 'update');

            endTest();
        } catch (e) {
            failTest(`Caught exception: "${e}"`);
        }
    });
    </script>
</head>
<body>
    <div>
        This tests the ability of the SourceBuffer to accept a new WebM Initialization Segment while still in the middle of parsing the Segment Element of the last Media Segment.
    </div>
    <video controls></video>
</body>
</html>
